debug:info od data availability is written to file correctly.

+++ nodes/ODOM_Data_Processor.py
@@ -1,6 +1,7 @@
 import numpy as np
 import rospy
 from pyquaternion import Quaternion
+import os
 
 
 class ODOMDataProcessor:
@@ -116,15 +117,19 @@ class ODOMDataProcessor:
             np.savez(f"{output_folder}/.odom.npz", coordinates=self.transformed_odom, saved_times=self.times_odom,
                      first_matrix=self.first_rotation_matrix_odom, first_transform=self.first_transform_odom,
                      matrices=self.matrices_odom)
-        with open(f"{output_folder}/.data_availability.txt", 'w+', encoding="utf-8") as file:
-            lines = file.readlines()
         is_odom_in_file = False
-        with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
-            for line in lines:
-                if line.startswith('odom'):
+        if os.path.exists(f"{output_folder}/.data_availability.txt"):
+            with open(f"{output_folder}/.data_availability.txt", 'r', encoding="utf-8") as file:
+                lines = file.readlines()
+            with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
+                for line in lines:
+                    if line.startswith('odom'):
+                        file.write(f"odom {state_odom}\n")
+                        is_odom_in_file = True
+                    else:
+                        file.write(line)
+                if not is_odom_in_file:
                     file.write(f"odom {state_odom}\n")
-                    is_odom_in_file = True
-                else:
-                    file.write(line)
-            if not is_odom_in_file:
+        else:
+            with open(f"{output_folder}/.data_availability.txt", 'w', encoding="utf-8") as file:
                 file.write(f"odom {state_odom}\n")
